Refactor components to different folder structure
