+++ app/home/textBlock.tsx
