+++ app/work/components/timelineDot.tsx
